+++ services/core/java/core-flow-service/src/test/java/org/ubiquia/core/flow/component/FlowEgressRelayTest.java
@@ -15,8 +15,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
@@ -36,7 +36,7 @@ import org.ubiquia.core.flow.service.registrar.FlowMessageRegistrar;
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
 public class FlowEgressRelayTest {
 
-    @MockBean
+    @MockitoBean
     private RestTemplate restTemplate;
 
     @Autowired