+++ services/core/java/core-flow-service/src/test/java/org/ubiquia/core/flow/service/cluster/ClusterSynchronizationServiceTest.java
@@ -25,9 +25,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.web.client.RestTemplate;
 import org.ubiquia.common.library.api.config.AgentConfig;
 import org.ubiquia.common.library.api.repository.AgentRepository;
@@ -51,10 +51,10 @@ import org.ubiquia.core.flow.service.cluster.synchronization.microweight.Microwe
 })
 public class ClusterSynchronizationServiceTest {
 
-    @MockBean
+    @MockitoBean
     private MicroweightClusterService microweightClusterService;
 
-    @MockBean
+    @MockitoBean
     private RestTemplate restTemplate;
 
     @Autowired